# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anatomiya",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/lesson": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "All Lessons",
                "description": "Hamma Dars(Lesson)larrni ko'rish",
                "operationId": "lesson_index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "new Lesson add",
                "description": "Yangi Lesson qoshish",
                "operationId": "lesson_store",
                "requestBody": {
                    "description": "Lesson Qo'shish",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title_uz",
                                    "title_ru",
                                    "title_en",
                                    "category_id",
                                    "images[]"
                                ],
                                "properties": {
                                    "title_uz": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "PHP"
                                    },
                                    "title_ru": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "ПХП"
                                    },
                                    "title_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "PHP"
                                    },
                                    "category_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "2"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "format": "number"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/lesson/{id}": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Show Lesson",
                "description": "bitta Lessonni hamma malumotlarini ko'rsatadi",
                "operationId": "lesson_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Update branch",
                "operationId": "lesson_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title_uz",
                                    "title_ru",
                                    "title_en",
                                    "category_id"
                                ],
                                "properties": {
                                    "title_uz": {
                                        "type": "string",
                                        "example": "title_uz"
                                    },
                                    "title_ru": {
                                        "type": "string",
                                        "example": "title_ru"
                                    },
                                    "title_en": {
                                        "type": "string",
                                        "example": "body"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "lesson_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "title": "ResponseError",
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "text",
                            "example": "Not Found"
                        }
                    }
                },
                "type": "object"
            },
...
            "Lesson": {
                "title": "Lesson",
                "required": [
                    "title_uz",
                    "title_ru",
                    "category_id"
                ],
                "properties": {
                    "title_uz": {
                        "type": "string"
                    },
                    "title_ru": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "student": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "bearer_token": {
            "type": "apiKey",
            "description": "Enter token in format (Bearer <token>)",
            "name": "Authorization",
            "in": "header"
        }
    }
}